+++ CHANGELOG.rst
@@ -3,11 +3,13 @@ Changelog
 
 Version 0.14-dev
 ----------------
-| **Date:** March 11, 2024
+| **Date:** June 20, 2024
 | **Release:** v0.14-dev
 
 Features
 ^^^^^^^^
+* Updated RC-index file to RC_1997-2024_June24_v5.dat (used during the
+  construction of CHAOS-7.18).
 * Increased precision of :func:`chaosmagpy.data_utils.timestamp` and
   :func:`chaosmagpy.data_utils.mjd2000` to nanosecond.
 * Added function to compute the unit base vector in the direction of the